servidor/requirements-api.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
mysql-connector-python>=9.0.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
